+++ ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: jcosta-b <jcosta-b@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/15 11:23:09 by jcosta-b          #+#    #+#             */
-/*   Updated: 2024/10/16 15:16:28 by jcosta-b         ###   ########.fr       */
+/*   Updated: 2024/10/17 16:01:35 by jcosta-b         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@
 
 void	*ft_memmove(void *dest, const void *src, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (ft_strlen(dest) <= n)
