+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: jcosta-b <jcosta-b@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/14 13:26:38 by jcosta-b          #+#    #+#             */
-/*   Updated: 2024/10/16 11:38:42 by jcosta-b         ###   ########.fr       */
+/*   Updated: 2024/10/17 16:01:47 by jcosta-b         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memset(void *dest, int c, size_t count)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < count)
